removes extraneous stuff from macros.Infrastructure
libraryClassLoader can be derived from currentClassPath

currentMacro can be trivially derived from macroApplication

Backend-detection methods forXXX (as in forJVM or forScaladoc)
might be useful, but current design of this API is not future-proof.
I'm not able to come up with a better design on the spot, so
let's remove this functionality for the moment.