Don't write public methods with non-public parameters.
If the parameter types of a method have lower visibility than
the method itself, then the method cannot be overridden because
the parameter types cannot be expressed.  This is a confusing
and frustrating situation to expose via public API.  Such
methods should either have access as strong as their parameter
types, or be made final.