SI-6215 Fix compiler crash on private method in value class
Fixes the problem with private defs in value classes by moving the $extension after the name proper rather than before. The previous scheme did not commute with makeNonPrivate:

I.e. if -ext-> is "generate extension name" and -mnp-> is "make not private" we did get for
method foo in value class Foo:

  foo -ext-> extension$foo -mnp-> Foo$$extension$foo

but

  foo -mnp-> Foo$$foo -ext-> extension$Foo$$foo

With the change both variations give the same name:

  foo -ext-> foo$extension -mnp-> Foo$$foo$extension

but

  foo -mnp-> Foo$$foo -ext-> Foo$$foo$extension