Recovered a bunch of deleted tests.
Are we in the habit of simply deleting tests when they
become inconvenient? A comment referenced test "0851" as the
example of why the code was needed; the test was deleted
years ago for no reason I can see except that it was not
passing at the time.  Words fail me.

Public Service Announcement: tests which are failing are
the MOST USEFUL tests.  DON'T DELETE THEM!