fixes a bug in a weak cache in runtime reflection
Entries in SynchronizedTypes.uniques could previously be garbage collected
in-between a successful call to contains and an actual cache lookup.

The patch could be a one-liner, but I don't want to use HOFs
in this function, whose prototype is a hotspot in the compiler.

Also the fix doesn't touch scalac in any way. It only applies to
reflective universes that provide runtime reflection functionality.