Worked over inferMethodAlternative.
We shouldn't let these core methods become so complicated. They are
way harder to follow than is necessary. It is why nobody ever fixes
the bugs in them. This is only the beginning really. So many things
in the compiler would border on trivial to fix if one didn't have to
navigate so much cruft and indirection.

Here are some methods which no longer exist:

 - hasExactlyNumParams. A dubious name for a method
   containing the expression "len <= n + 1".

 - resolveOverloadedMethod. A method which returns a list
   of symbols can't be resolving all that much.

 - isUnitForVarArgs. Take a guess as to what that method does.
   Ha ha, you were not even close.

Still on my hit list:

 - "very similar logic to doTypedApply in typechecker"
   I find a good rule of thumb is never to write a comment
   which paraphrases to "this is very similar to that."

This entire patch is the fault of a. moors for trying to
cherry-pick a comment of mine from github into trunk. This
patch hopefully makes the comment unnecessary.