Moved a bunch of passing tests out of pending.
If the test names can be believed, this covers

  SI-294 SI-1751 SI-1782 SI-2318 SI-3897
  SI-4649 SI-4786 SI-5293 SI-5399 SI-5418
  SI-5606 SI-5610 SI-5639

Most of these were moved to pending in 1729b26500
due to failures of unknown cause.  It was suggested
they be brought back "as soon as possible" and that
was three months ago; I suppose it's now possible.

If they need to be disabled again, please move them
to test/disabled, not to test/pending.  "disabled"
should mean a formerly passing test in limbo; "pending"
tests document bugs which await fixing.

I also removed some dead files in test/ - the
files with a "cmds" extension are from a failed
experiment and do not do anything.