Fix for SI-6537, inaccurate unchecked warning.
I found a more direct expression of the unchecked logic,
which should be much easier for others to verify.  But the
bug being fixed here is that the unchecked checking happens
too early, and the sealed children of a symbol are not yet
visible if it is being simultaneously compiled.