Introduces some structure for name lookups.
Too many unrelated things are intermingled.  If you
want to know what symbol corresponds to a particular name
in a particular context, you shouldn't have to involve
either a Tree or a Typer to find that out.  I toiled line
by line over typedIdent until it had shed its redundancies
and freed itself from the bowels of typed1.

The mechanism of name lookup is such that adding a
qualifier when the occasion calls for it is inseperable
without a lot more effort.  So to preserve a sane interface
I devised this small partitioning of outcomes.

  case class LookupSucceeded(qualifier, symbol)
  case class LookupAmbiguous(msg)
  case class LookupInaccessible(symbol, msg)
  case class LookupNotFound(msg)