Make isOverridingSymbol much cheaper.
Use the OVERRIDE flag as a cache for positive results.
This dropped over 40% of the calls to matchingSymbol in
a compile of trunk.

  Before: 3541107
   After: 2067403

I think most of the rest of the calls would disappear as well, if
I had some way to cache negative results. Unfortunately our usage
of flags is too ad hoc to allow for that. There is even an ironic,
tauntingly named flag known as "notOVERRIDE" which is not usable for
marking a symbol as not an override.

If I ran the world, OVERRIDE would be something set by code (not
by humans) at symbol creation time and either never modified, or
modified at the fewest possible well-defined times. There's no need
to make every not-overriding symbol double-check with every ancestor
every single time, on the off chance one has sprouted a matching
method. In almost every case the override status of a symbol at
creation will be its override status forever.

I stole this optimization idea from retronym in #1074.