Objects are now consistently called objects, not modules.
The good thing is that even if we choose a different name,
this commit serves as a more or less precise documentation
of places we need to touch to fix the API.

This also solves possible confusion between Modifiers and
Modules, which used to have nearly the same identifiers in code.

Popular members in the reflection API keep the old names as
a @deprecated member.

MODULE$ and other significant binary compatibility artifacts
were left intact.