Fixes SI-6500 by making erasure more regular.
With the introduction of value classes, erasure uses specialErasure where a value class C with underlying type T is unboxed to an ErasedValueType. ErasedValue types are eliminated on phase later, in post-erasure. This was done everywhere, except in the parameter types of bridge methods. That was a mistale, because that way  bridge methods could not do the boxing/unboxing logic triggered by ErasedValueTypes.

Note: there is one remaining use of erasure (not specialErasure) in Erasure.scala. I put in a comment why that is OK.