Fixed bug in Symbol filtering.
If you called filter on an overloaded symbol, it tried to
return itself if no alternatives were filtered out. The test
being performed, however, would only ever be true if the
list call was to the (non-existent) "filterConserve", which
is to say that in general,

  xs ne xs.filter(_ => true)

The upshot is that we were creating a new symbol on every filter
call to an overloaded symbol. To make completely sure this would
be a performance winner, I also eliminated the closure and perform
the filtering locally.