Update tools/epfl-publish
Currently, during each Scala nightly build, the epfl-publish script rsyncs the resulting dists/archives to the repository machine (chara).

Now, as a result of a change introduced a while back, dists/archives/ now also contains a symbolic link, created at the end of the build when the targets "pack-archives.latest.*" are invoked (in src/build/pack.xml). That was introduced in:

scala/scala@506bcfe71c93160ebfa0ca9b8b170b4b54e844e9
scala/scala@cb99853c8655686dae1288cbcd44a42cf1ea6609

This link, created in dists/archives/, is copied over as-is at the end of the rsync to chara by epfl-publish. On chara, however, the link points to an invalid target (the path is absolute).

Separately, the repository directory on chara is rsync'd every 30 minutes over to the machine that serves www.scala-lang.org, via cron. Rsynch finds that the link target does not exist, and generates an email with the crontab log, which is then duly sent to a system mailbox. Every 30 minutes. Since March. Needless to say, the mailbox is pretty large by now.

The fix is trivial, but this needs to be included in all the branches that are tested during the nightly. Since RC2 has just been cut, and the code frozen, I'm not sure how this change can be included without disruption; therefore, I'll just send a pull request on 2.10.0-wip, and let Josh/Paul take the appropriate steps.