SI-6441 Cull scala.swing.SwingWorker
It appears to do nothing much at all, refers to the now-deprecated
scala.actor package, and has a name clash with an important class
from java.swing.

Note that the POM for scala-swing does not declare the dependency.

The final nail in the coffin is it aides and abets the crash in
the crash in SI-6440.

Obviously removal of a class isn't binary compatible with 2.10.0.
Things could play out a few ways:

  - RC3 is required for some other reason, and this change is
    admitted to 2.10.0
  - Or, an exeption to the binary/source compatibility policy
    is made for this class
  - Or, this is deferred until 2.11.0, or at least until the
    spin-off of scala-swing to a standalone repo.

For now, I'll submit this to 2.10.x for now and leave those
choices up to the release managers.