SI-6641 Cull scala.swing.SwingWorker
It appears to do nothing much at all, refers to the now-deprecated
scala.actor package, and has a name clash with an important class
from java.swing.

Note that the POM for scala-swing does not declare the dependency.

The final nail in the coffin is it aides and abets the crash in
the crash in SI-6440.

Attempts to remove or even deprecate this in 2.10.0 were
missed the boat.

This commit removes them for 2.11.0, in contravention of
the deprecation policy. Given that they rely on the
scala-actors module which _is_ deprecated in 2.10.0,
and will likely be removed in 2.11.0, I contend that
this is a reasonable exception.
Actor.scala