Added -Xdev setting... you know, for devs
A setting we developers can give all the time and expect to
hear useful things without being buried in debugging output.
As the comment says:

  This is for WARNINGS which should reach the ears of scala
  developers whenever they occur, but are not useful for normal
  users. They should be precise, explanatory, and infrequent. Please
  don't use this as a logging mechanism. !!! is prefixed to all
  messages issued via this route to make them visually distinct.

This is what I always intended for "debugwarn", the method
I have deprecated in favor of the more accurate:

  def devWarning(msg: => String): Unit

In this VERY SAME COMMIT, I performed the CLOSELY RELATED
task of quieting down an -Xlint warning which had become too
noisy thanks to implicit classes tickling it. I tightened that
warn condition to include both -Xlint and -Xdev.