Turn classOf into a macro.
A small victory for the cause of macros removing special
cases. classOf[T] now implemented with a macro in Predef,
with no further special treatment required.

There is a minor behavior change on an unlikely corner case
(SI-4871) which for the sake of consistency should work as
it does now, which is to say, not work.