SI-6415 avoid early evaluation in some Stream's methods
The lengthCompare method in LinearSeqOptimized was looking one step further
than it needed to in order to give the correct result, which was creating some
unwanted side effects related to Streams.

Add test case to expose the problem and a few test cases to ensure correctness of LinearSeq.lengthCompare() method.

https://issues.scala-lang.org/browse/SI-6415