SI-6745 Skip refinements in the search for <init>
These arise in the base class sequence when a self type
is annotated. This prevented an auxillary constructor from
calling the primary constructor.

Unlike the first version of this change, here we do this
`typedIdent`, where we can easily determine the enclosing class.

Also closes SI-4460, which showed a more insidious version of
this bug that compiled with invalid bytecode.