Javap for repl output.
The enabling fix is in AbstractFileClassLoader: the javap tool
uses a ScalaClassLoader to pick up tools.jar, but delegation to
the parent AbstractFileCL was broken; minimal support for findResource
was added.

Other minor tweaks to output: retain the weird "Binary file...
contains" message as documentation; the message is not output for
:javap $line3/$read$$iw$$iw$Foo.class, for example.

No option handling by javap-7 except to catch IllegalArg, so that
combined arg style is not yet supported (:javap -pv).