SI-6751 macroContext factory is now public
In order to make macro contexts customizable by the embedding environment
such as an IDE.

The motivation for this commit was a request to expose an IDE-friendly
debug log API for macros. With macroContext being overridable (see the
instructions in the comments), we can recommend people to use c.info
and then the IDE could override it and present it conveniently.