Restore unmangling but add -raw; massage options to support tool args like -raw and combined -pv, and not pass through arbitrary options.
If class arg fails, try it as a live term and use its
enclosing class, to support "def m = 7; javap m". That
might be more fun with filtering (to see only m). This
saves having to know the line and iw's. There are no
doubt more fun features for another rainy day. And it
will rain again someday.