Support "javap -" for lastvar, move JavapTool into JavapClass.
The notation "javap -" is actually pretty handy, in particular
as a shorthand for javap resN.

JavapClass knows if it has an intp, for purposes of names and
for output; so the special tryClass impl is refactored to the
tool; the special handling for "-" as well. It just uses IMain
for withoutUnwrapping, so that method is unadded from ISettings.

(The tryClass/tryFile template methods are handy but maybe
passing a func to JavapClass.apply would be cleaner; similar
for writer output and whatever magic it enables.)