Fix for SI-5717, crash in GenASM.
If there is a file in the way when generating classfiles,
remove it and proceed. If you had wanted to keep it, you
wouldn't have left it lying around in an output directory
with a name which stands between you and the filesystem.

This is too difficult to write a test for under partest.